Fixing Bugs

+++ app/Http/Controllers/GuruController.php
@@ -30,6 +30,7 @@ class GuruController extends Controller
                 ->addcolumn('usernames', function ($guru) {
                     return $guru->user->email;
                 })
+                
                 ->addcolumn('button', function ($guru) {
                      return '
     <div class="flex justify-center space-x-2">
@@ -39,7 +40,7 @@ class GuruController extends Controller
             </svg>
         </button>
         <button class="text-red-600 hover:text-red-900" title="Hapus"
-                onclick="openConfirmModal(\'' . $guru->ID_Guru . '_' . $guru->user->username . '\')">
+                 onclick="window.location=\'' . url('/guru/' . $guru->ID_Guru . '/' . $guru->user->username . '/destroy') . '\'">
             <svg xmlns="http://www.w3.org/2000/svg" class="h-5 w-5" viewBox="0 0 20 20" fill="currentColor">
                 <path fill-rule="evenodd"
                     d="M9 2a1 1 0 00-.894.553L7.382 4H4a1 1 0 000 2v10a2 2 0 002 2h8a2 2 0 002-2V6a1 1 0 100-2h-3.382l-.724-1.447A1 1 0 0011 2H9zM7 8a1 1 0 012 0v6a1 1 0 11-2 0V8zm5-1a1 1 0 00-1 1v6a1 1 0 102 0V8a1 1 0 00-1-1z"
@@ -49,6 +50,9 @@ class GuruController extends Controller
     </div>
 ';  
                 })
+                 ->editColumn('created_at', function ($row) {
+                return $row->created_at->format('Y-m-d H:i'); 
+            })
 
 
                 ->rawcolumns(['username', 'usernames', 'button'])
@@ -203,10 +207,18 @@ class GuruController extends Controller
      */
     public function destroy(string $id, string $id_user)
     {
-        $guru = guru::find($id);
+        $guru = guru::find($id); 
         $user = User::find($id_user);
-        $guru->delete();
-        $user->delete();
+
+        if ($guru) {
+            $guru->delete();
+        } else {
+            return back()->with('error', 'Data guru tidak ditemukan');
+        }
+
+        if ($user) {
+            $user->delete();
+        }
 
 
         return redirect(route('get.guru'))->with('success', 'Guru berhasil dihapus');
